Remove dynamic serialization switching
The code for dynamic serialization switching can be tricked into
behaving incorrectly. For instance, a request for .../1.1.2 will
appear to be a request to serialize to troff (".2"). This is not the
desired behavior.

The dynamic serialization switching feature should be re-implemented
at another level of the stack, or should use a different separation
character, e.g. '?'.

Fixes: bug #1197479
Change-Id: I4610b620f52f5eefad65dcbeaa3114bc43547ba2