Neutron support
Added configurations:
* use_neutron

If Neutron is enabled, management network id is
passed as a parameter, when instances are being
launched.

Instance's internal and management IPs are equal and
point to the first IP in a list of a specified
management network.

Model classes for Cluster and Cluster Template updated.

Implements blueprint add-neutron-support

Change-Id: If6e79327ab14bbc813175396cd7287c2a29149b0