Replacement of Vanilla Hadoop 1.1.2 to Hadoop 1.2.1
This patch is needed for EDP purposes and will keep
Vanilla Plugin with the latest stable Hadoop as well

In Hadoop 1.2.x version,  4 level topologies were introduced:
all network, rack, node group (represent Hadoop nodes on the
same compute host in the simplest case) and node.
Usually Hadoop has replication factor 3. In this case
Hadoop placement algorithm is trying to put a HDFS block
in the local node or local node group, second replica
should be placed outside the node group, but on the same rack,
and the last replica outside the initial rack.
Topology is defined by the path to vm e.g.

/datacenter1/rack1/host1/vm1
/datacenter1/rack1/host1/vm2
/datacenter1/rack1/host2/vm1
/datacenter1/rack1/host2/vm2
/datacenter1/rack2/host3/vm1
/datacenter1/rack2/host3/vm2
....

Also, this information will be used for job routing, to place
the mapper as closest as possible to the data.

Implements blueprint vanilla-plugin-hadoop-1-2-1

Change-Id: I39e11c8d59f6eae317e8ebe989c113aec413b33e