Filter out some vendor based input from a template upload
Some values in the vendor blueprint (e.g. Ambari blueprint) are not
intended to be displayed to users.  They are now filtered out at
the convert() call to prevent issues during _map_to_user_inputs
(it makes use of get_configs() and can raise exceptions is the
input is not returned).

Fixes: bug #1226109

Change-Id: Iae6afa65221f9ef7c5025c9f8d9dbc92f32da7bb