Fix exception handling in Savanna subprocessing
Some exceptions could not be serialized by pickle in python 2 because
of the following bug: http://bugs.python.org/issue1692335

Hence current exception handling logic is Savanna subprocessing is
broken: If an exception occurs in child process, it is pickled and
passed to the main process. Then it is unpickled and thrown in the
main process.

The fix changes the behaviour in the following way:
 * Exception message is passed instead of the exception object
 * In the main process SubprocessException is thrown with the message
   passed from the child.

Change-Id: Ib8fac65d2df93cbb585bef02df7576aa6d0a81c9