Fix the _assert_types test to allow for fields that are enums
Fields marked with "type: string, enum: [choices]" in JSON schemas
fail the _assert_types test because the error message is different.
When an enum field is type checked against an int, the message
becomes "<int> is not one [choices]".  This change checks for a
different message when an enum is specified.

Change-Id: I11f4f652e1522e06180d1cf537f148fd18403975