Enable cluster deployment with pre-installed JDK
This commit is the first one in a series to support cluster deployment
for VMs with no external access (e.g. access to package repositories).
This commit supports the ability to specify to Ambari that a JDK already
exists in the specified location.  Therfore, the JDK download from a
potentially external location will be prevented and the existing JDK
pre-installed on the VM image will be leveraged.

The full disconnected facility still depends on the following (which will
be addressed in subsequent commits):

- additional testing
- an updated HDP image with the full set of required packages
- a fix to an ambari issue with nagios installation discovered
during testing

partially implements blueprint hdp-specify-local-repo

Change-Id: I36fb109225c6ea21114d7e80e7d3ce950ad0b6a7