Oslo sync: make wait/stop funs work on all threads
Previously, ThreadGroup.wait and ThreadGroup.stop would iterate over
the self.threads list and call the wait or stop method on each one.
This list could potentially be edited concurrently by the _thread_done
callbacks and remove some items in the list.  As a result, the wait
method would not actually wait for all threads in the thread group.

Additionally, vim headers was removed.

Change-Id: I254992b06c24cbf31f02e8f2b29430df57c85da6