Eliminate extra newlines in generated workflow.xml
When the base workflow in service/edp/resources/workflow.xml is read
newlines are preserved in the text elements.  Additional elements are
added without newlines and newlines are inserted when the text is
generated.  This causes extra newlines after the original elements.

This change strips whitespace from the beginning and end of each
line in the base workflow file when it is first read, which eliminates
the extra newlines in the text elements.

Closes-Bug: #1265916

Change-Id: Ibf75cdf54c3218df9d0085005624b3ee8d4c0ea5