Read Swift credentials from input_data OR output_data
Fix the workflow generation so that Swift username and password
configuration values can be automatically generated from the
input data source or the output data source.  Previously, only
the input data source was considered.  With the addition of
other storage technologies for data sources, it is possible that
only one or the other is in Swift.

Closes-Bug: #1270201

Change-Id: I61434fe658dd6bf30133947f910cfca00fe0e6ca