Change configs["args"] to be a list for Pig jobs
The schema for Pig jobs incorrectly listed "args" as a dictionary.
It should be a list of strings for all job types. Modify the schema
and workflow generation to correctly handle "args" as a list.

There is a UI component to this bug as well.  To avoid having to
sync changes in savanna and savanna-dashboard, temporarily allow
"args" to be passed from the UI as a dictionary and replace it
with an empty list. This will prevent the UI from breaking.

Partial-bug: #1269968

Change-Id: Ieacb0596dbb6da5d5f26dc6a5b923f2bafa819f2