Allow boolean "streaming" in Job JSON
Allow a boolean "streaming" to be set in the POST of a job object.
This flag will loosen validation for Job objects when the type is
MapReduce, so that "libs" is not required.  This is in preparation
for streaming MapReduce jobs.  It is expected that the UI will
set the "streaming" flag for MapReduce based on user input.

Implements: blueprint edp-oozie-streaming-mapreduce
Change-Id: I9f35e3de53e5c9c3234f9b20e7f051e146b5a5ad