Enable HDP 2 deployment leveraging HDP plugin
This commit introduces the ability to deploy HDP 2.0.6 clusters.
Both basic cluster provisioning functions (create, scale) as well
as EDP have been implemented.

There is some follow on work required to complete the integration:

- Topology/data locality support
- Verify cinder support
- Refactoring (there is ample opportunity for extracting parent
classes for common code, etc)

I have filed bug #1277653 to track these items.

Most of the items above were not undertaken as part of this commit in
an attempt to limit the scope/size of the commit as well to allow for
an iterative approach for achieving full functionality.

partially implements blueprint hdp-hdp2-support

Change-Id: I1cc483f05b65e1df0fb630d5e819ef23ab407492