Add support for dotted job types
This change supports dotted job types.  A job type consists of a type
followed optionally by a '.' and a subtype.  The first dotted subtype
Savanna supports is MapReduce.Streaming.  The Streaming subtype
distinguishes it from a regular MapReduce job and indicates a
difference in the Oozie workflow generation and required configuration
parameters.  The goal is to maintain parallelism with Oozie action
types where Ooozie treats some features as a variation of a top-level
action.

Partial-implements: blueprint edp-dotted-job-types
Change-Id: If2ebd866825980c9c03feda524973b14beb89471