Fixed hadoop dir creation during hadoop-swift lib download
During installation of IDH 3.0.2 there is no /usr/lib/hadoop/lib
folder at moment of the code excecution. "curl --create-dirs" ignores
umask and creates folder with 750 permissions instead of expected 755.
And therefore manager fails to format HDFS later.

Installation of hadoop-swift need to be revised in the future, but for
now this patch fixes the issue.

Applying the patch to IDH 2.5.1 to get rid of dangerous
"curl --create-dir"

Change-Id: Icbc781f6a36ababb362a456e2bf46bea871e546c