Expand cluster-template usage validation message
With the introduction of a CLI, users are presented directly with the
error messages sent from the API service. Specifically, when cluster
templates are delete a usage check validates that there are no users
of the template. This commit expands the message returned from the
check to include clusters where the template is in use.

Change-Id: I516d107edc93e24832ccb952f3e68defba92adcd