Filter 'fields' from JobExecutions returned from REST api
The 'job_configs' field will contain swift credentials used by Hadoop
when running jobs that use swift data sources.  The 'extra' field may
contain a token for use with neutron.

The current filtering mechanism in Sahara allows filtering fields by name,
but it's not set up to support programmatic filtering of things which
can be variable. It may not be necessary to filter out 'job_configs'
entirely from a security perspective, but there currently is no other
option when a field might or might not contain particular values.

Additionally, sensitive information could potentially be passed in 'args'
within 'job_configs' and it is impossible to know anything about the
content of that field.

Closes-Bug: #1273661
Change-Id: Idb2e68a2d42e45bab04c62c740cbbaf5e51b2719