Add sahara-all binary
In the nearest future we are going to split current sahara-api process into
two: sahara-api and sahara-engine. The new process sahara-all will replace
the old sahara-api in that it will run standalone Sahara all-in-one, which
does not require running several different processes.

This patch set will help us make a painless transition from sahara-api
to sahara-all in our CI. The plan is:
 * merge current patch set, which just adds sahara-all alias to
   sahara-api
 * change CI to run sahara-all instead of sahara-api
 * merge patch set which splits sahara-api and sahara-engine

Change-Id: I4fa68364b756c803a3fa8a0b92813785636ea13c