Allow HDFS data source paths without the hdfs:// scheme
HDFS paths without a leading hdfs:// scheme will be interpreted by
hadoop to refer to the local hdfs.  Relative paths will be treated
as relative to the user running the job; absolute paths will start
at "/" in the local hdfs.  Sahara should allow these forms for
paths so that data sources can simply reference the local hdfs
on a long running cluster.

This can be tested/utilized from the CLI, but the UI will need
an additional change to allow a user to submit these paths.

Partial-Bug: #1315126
Change-Id: I575bf9ff20ff348f7bf2bc52f116b74288b8464a