Fix hardcoded tenant name for job binaries
After the fix the current user's tenant is taken instead of one
from the Sahara config.

I think we can fix it better and I outlined my thoughts in the bug
comments. Current fix could be viewed as temporal until we agree on
something else. BTW it could be ported to stable/icehouse since it
does not changes Sahara API, unlike the fix I've proposed in the
comments.

Change-Id: Ifef76905cab313e451eb9e68f915738f33e11a25
Partial-Bug: #1306713