Remove usage of remote from HDP Instance constructor
When Sahara is run in distributed mode, the Instance object is used in
sahara-api context. You can see the full stack trace in the bug description.
Here it is in the simplified form:
  * User requests HDP cluster scaling
  * Scaling validation (in HDP plugin)
  * HDP Instance objects are constructed
  * Sahara's instance.remote() is called inside the constructor

The remote is unavalable in sahara-api context and that causes the issue.
Moving invokation of instance.remote() out of constructor fixes it.

Change-Id: I0d09c7a19e4df7fce948994fb21028fe147d2782
Closes-bug: #1325989