Add Spark 1.0.0 to the version list
We tested an image with Spark 1.0.0 created with the DIB and it worked
without need for any other change in the plugin.
As soon as the documentation patches are merged, I'll submit an update
for those files as well.
Version 0.9.1 is not removed from the list since there are already
images floating around and people are using it.

Change-Id: Ie41d9fbbd340fc9527fd5e8b0fc48eed0f6705ad