Fixed cluster rollback on scaling with heat engine
We can't allow stack to move into Error state if we want to use it futher.
disable_rollback parameter controlls such behavior. It is Ok to have errored
stack for cluster create and rollback since we will not be using it anyway. But
for scaling we need to keep it Active since we can just rollback in case of
error.

Change-Id: Iab1a58f4924ca6a7689060f71217956a72316517
Closes-Bug: #1337006