Adding job execution examples to UI user guide
Adding examples for a Pig job and a Spark job to
the Sahara UI user guide.  The examples referenced
are the ones from the Sahara repository /etc/edp-examples
directory.  Also, removed language that referred to the
"Sahara" dashboard that formerly existed as a separate
entity.  The updated version references the "Data Processing"
panel where it now lives (under Project).

Change-Id: If83cdef109a4356fdb2db4487cde4e76d7d280a8
Closes-Bug: #1288914