Removing Swift container support for job binaries
This change removes the ability to request an entire container as a
source for a job binary. The reason for this removal is that the
internal code paths for retrieving job binary raw data from Swift will
need to be refactored to support this functionality. The scope of the
changes needed will impact the REST API and the concepts surrounding how
job binaries are aquired. Allowing full containers should be discussed
in a new specification to add the feature.

Changes
* removing container support from get_raw_data.
* moving the functions to acquire a Swift client into a more appropriate
  place.
* removing context argument from internal_swift.get_raw_data
* creating tests for internal swift get_raw_data function

Change-Id: I666873038a323e63d17b8a24b4213ed0240a4c59
Closes-Bug: #1306505