Support searching job executions by cluster name and job name
As a convenience, allow job execution searches to be filtered by
'cluster.name' and 'job.name' fields in addition to the fields
in a JobExecution.  These fields refer to the names of the Cluster
object and Job object linked to by the  JobExecution. If these
search terms are present in the JobExecution lookup, SQL joins
will be added to accomplish the search.

Partial-implements: blueprint enable-result-filtering
Change-Id: Ibdffa81c6e15e3863d40bdb544721defea1d41fd