Include launch_command.py in MANIFEST.in
The file launch_command.py was missed in the MANIFEST.in for Juno,
which results in a non-functional Spark cluster.

See https://bugzilla.redhat.com/show_bug.cgi?id=1184522 for one
expression of this bug.

The change to fix this bug was rolled into this review:
https://review.openstack.org/#/c/146659/13
It is not appropriate for cherry-picking, as the change with
which it was made is much larger, but nonetheless this issue does
render Spark clusters inoperable in Juno. The change is approved
in master. Please let me know if a more formal process of
registering the bug and attempting to create an atomic commit is
needed; it seemed to me that given the state of this issue,
attempting to do so might only create a more tangled web, but I'm
happy to do so if it is preferable.

Closes-bug: #1419868
Change-Id: I29a17eb3ae9f9a3b8c5412d0be1813cfeca99622