Release Import of Translations from Transifex
Manual import of Translations from Transifex. This change also removes
all po files that are less than 66 per cent translated since such
partially translated files will not help users.

This updates also recreates all pot (translation source files) to
reflect the state of the repository.

This change needs to be done manually since the automatic import does
not handle the proposed branches and we need to sync with latest
translations.

Note: This is part of importing of translations, there are no new
translations for this project, thus only the pot file gets updated.

Change-Id: Ia187bb71df998908c40281f71003d7cf3d092dd2