Add spaces around function params for browser to linewrap on
The modified function names with parameter do not wrap on the webpage,
due to not having a normal breakpoint (usually spaces).  Including a
space after the opening paren, as well as one before the closing paren
to visually balance, seems to be the best option I can see for fixing
this while still maintaining readability and coherence in the rst file
itself.

Closes-Bug: #1373110
Change-Id: Iea09ca21ff50a51bb28f518b0b124f129cf06ae2