Allow Mako templates as input for scenario test runner
The test runner can read Mako templates, dynamically replacing the variables.
The values of the used variables must be specified in a INI-style file, which
is passed as argument (-V) to the test runner. Please note that a missing or
incomplete INI file does not block the execution, as it is not known which
variables are needed by the specified templates beforehand.

The runner unit tests have been expanded as well.

Partially Implements: bp scenario-test-config-template
Change-Id: I7d19beb82f5b5b6ae93c82768a1ea73c362ca01d