Allow Sahara native urls and runtime urls to differ for datasources
When we consider things like Manila nfs shares as datasources,
the possibility arises that the native form of a datasource url
in Sahara may not be the same as the form of the url needed at
runtime in the cluster. Allow them to differ, so that we can
still record the Sahara native url form in JobExecution objects
for accurate reference while passing the correct runtime url
in job arguments, etc.

This is a base CR that will be further built on later. In this
change, native urls and runtime urls are always identical.

Change-Id: I53f4cf11320e112ffd0c4ae93b7d1f300df86878
Partial-implements: blueprint manila-as-a-data-source