Changing scenario runner to use subprocess
The os.system call returns an exit status on Unix systems that is
similar to the os.wait command. This format is a 16-bit number whose
lower byte is the signal number that killed the process and the higher
byte is the exit status. This caused errors in the way that the scenario
runner script returned the exit status of the scenario test.

To fix this issue, the runner has been modified to use subprocess.call
in favor of os.system. This solution should be portable across operating
systems while maintaining the expected functionality.

Change-Id: Iea95312702977a5f2a194397abf214ae20c2e23d
Closes-Bug: 1486493