implement is_sahara_enabled
The is_service_enabled method in devstack requires is_sahara_enabled
to be defined somewhere to be used in additional selectors like for
tempest configuration and possibly upgrade work later.

While trying to build a composit job that ran trove / heat / sahara
tests all together this was found to be missing, so sahara never got
configured for tempest.

Change-Id: I69acf6ab2d74d02e9fe8e2d419e54c9367fca500