remove hdp from the default plugin list
as per the irc meeting on february 26th 2016, the hdp plugin is being
removed from the default list. this decision is mainly based on the
ideas that hdp will be deprecated in the newton cycle, and that it has
not received updates after version 2.0.6.

this also fixes the unit test for default plugins by adding all the
entries from the configuration option in the base plugin module.

Change-Id: Iaa9bc4cc9fefeba6ece9ceee9605b1fbf2df9513
Closes-Bug: #1549965