Added #nosec to sahara.service.coordinator package
The sahara.service.coordinator module uses an MD5 hash to identify
object instances that it knows about. These object identifiers are
not exposed outside the application, and as such have a low
likelihood of generating hash collisions produced by an attacker.
The usage should be marked with nosec as per the bandit documentation.

Change-Id: I995ae9eedc35f387c2259d62c25e5d73252de683
Closes-Bug: #1552476