get_admin_context overwriting context
When the admin context is retrieved, the user context and its request
ID is lost. All subsequent log entries would have different request IDs
and any further context checking will be using an admin's context
instead of the user's context. I'm not aware of any further checking
today, but better to fix this now and avoid that being an issue in the
future.

Change-Id: I1371ddb70736a49f9124f85352d6d195aec97c80
Closes-Bug: #1511406