Error handling during hosts file generation
Cluster deployment currently fails if hostname of
service endpoint cannot be resolved. An error like
this should not be critical since it could be the
result of a temporary DNS issue, etc. The patch will
allow deployment to continue. A warning will appear
in the logs and in the hosts file of each instance.

Change-Id: I8e4d88ec782dec900b39135e49da985bf9d4f39b
Closes-Bug: #1616112