Remove html_static_path from api-ref
In Id3a71841f5b9a9ea7bb5803afd967ab4feadfe10 the in tree copy of
os-api-ref was migrated to the upstream version. The html_static_path
setting in conf.py was forgotten in the patch, which caused a warning
when building the API reference.
This patch comments out the setting from conf.py to prevent the
warning.

TrivialFix

Change-Id: I595a5a85db9fbacdf17f1bc4706ee4cc6705d81a