Replace 'sudo pip install' with pip_install
Configuration of sahara in devstack behind a proxy fails because the
installation of "tooz[zookeeker]" does not work.

This is because sudo does not preserve environment variables, so pip
does not use the configured proxy. Using 'pip_install' instead of
'sudo pip install' should fix this, since 'pip_install' ensures that
the environment is preserved.

Change-Id: I78d483e86c586846f92bdc7118d495efe7e6df17
Closes-Bug: #1632673