Use http_proxy_to_wsgi middleware
This sets up the HTTPProxyToWSGI middleware in front of Sahara.
The purpose of this middleware is to set up the request URL
correctly in case there is a proxy (For instance, a loadbalancer
such as HAProxy) in front of Sahara.

The HTTPProxyToWSGI is off by default and needs to be enabled
via a configuration value.

Change-Id: Ica7e8671e3880c0db90d382bec89b0994f75b36d
Closes-bug: #1590608