Fixing manila microversion setting in sahara.conf
When setting manila microversion in sahara.conf, there is
an exception: api_version is not valid: invalid literal for
int() with base 10.Reason is the type of api_version is set to int in
sahara/utils/openstack/manila.py - Changing it to string fixes it.
and then we can set microversion like 2.2 in the sahara.conf.
Closes-Bug:#1664825

Change-Id: I588dd93a0ee2efb3f311a5e7d42a8a03176f877c