Optimize model relationships (avoid joins, prefer subquery)
There are a large number of unnecessary joins which are currently
duplicating data and increasing the number of rows for simple
SQL queries.

It is not very easy to add the lazyloading again without a big
change in the architecture due to the dependence of these
components.  For those relationships that require to be there,
they will use `subquery` which will send much less data over
the line.  For the relationships that can be lazyloaded,
it was dropped.

Change-Id: Iddaaf7168611763a6fa76818e3be9d1aa24d8021
(cherry picked from commit e20d0bee721d93d8a2de7c22582e27af95c93386)