Fix instances schema doesn't sync with nova instance
Create a cluster more than 10 slave nodes, or create a cluster
named against the first half of the alphabet and scale up it
after renaming the cluster named against the second half, and
then scale down the cluster. Cluster will go to Error
"No router to host".

The root cause is that the record in instances schema doesn't
sync with nova instances. Heat deletes instances according to
the index of instances, so sahara should keep with heat to delete
the same on.

Change-Id: I05182a7f7b37a66caf45d56716f7b84a93c5c7f2
Closes-Bug: #1713012