Use default log levels overriding Sahara-specific only
Do not override all default logging values provided by oslo_log, but
only the levels required by Sahara.
The default values contain for example the defaults for 'amqp' and
'oslo.messaging', which were not covered (and fills quickly the logs).
Instead of manually synchronizing the default values (a process that
can be easily forgotten), just use whatever the defaults provide.

Closes-Bug: #1741233
Change-Id: I0ab6993e82f895e15d0367e3a404b3751fc3f36a