Fix Flask error_handler_spec
The error_handler_spec is supposed to be a dict of dict of dicts, but
we had it as a dict of dict of functions, for unknown (historical?)
reasons.

Fix that, so that errors which occur in the Sahara API, but are not
otherwise caught by `sahara.utils.api.Rest.route` get handled correctly
by Flask instead of leading to some arcane error far down the stack.

Change-Id: I1e9d5f3fa00308baa7eee101c1f3b5a666cae4aa
Closes-Bug: #1745236