Remove step upload package to oozie/sharelib
It is not robust to upload a special package in a common method.
However, because of change from Hadoop(2.8.2 and later)side,
we do need the dependency package "commons-httpclient" in
oozie/sharelib to let oozie work.
This patch removes the uploading step in sahara because there is
a patch 2eedcec728b4f93e4cb226c5137159dcb3ddbfa3 which add
"commons-httpclient" as a dependency to oozie/sharelib.
So when we build the oozie binary package, "commons-httpclient"
will be as a part of it.

Change-Id: I4f20a356efc2dcc51aac2f10105be6aa351f87e8