Migrate the artifact link to sahara-extra, use https
The stable link for artifacts is going to be tarballs.openstack.org/sahara-extra,
which matches the name of the repository and it is more consistent
with the way the publishing system works.

Also use https.

Change-Id: Iacce6e6397b2d8d812964d56a33f9ea82901a3d6
(cherry picked from commit 8ae30c08fa79559af33c3c3334c21e6fce34bc8e)