Fix the installation of the Swift Hadoop connector (CDH)
The jar file with the same name already exists because
it is installed with the rest of the Cloudera packages,
but it must be overwritten by our version with Swift support.
So do not exit if the jar already exists.

Story: 2001909
Task: 14422
Change-Id: Ibc0a5f7348365f2a6c5658846ee74aa92e78314b
(cherry picked from commit 8cc87685cc63a38050da620884f3077ef364bac0)