Check node processes earlier
When creating a cluster, validate node processes as early as we can.

(The config-recommending step may have some impact on validity, so we
must leave that as the first call.)

Note that there has been a desire mentioned to move the node process
validation to occur during cluster template creation. Unfortunately the
amount of refactoring needed makes the task too daunting to be
completed now.

Change-Id: Ib5c91e062b32a83268d178417cbc5120d4c57934