Fix flask.request.content_length is None
In Python 3 (I haven't checked with Python 2), it appears that sahara-api
is broken, checking for flask.request.content_length > 0, when this
object doesn't exist, and then it fails with:

TypeError: unorderable types: NoneType() > int()

This is due to differences between py2 and py3.

This patch therefore tests if flask.request.content_length is None in the
validation code.

Change-Id: Iedddc1b2a8d5c3d88d02a380eaf85206a234494a