doc: light cleanup of the ironic-integration page
Remove the commands required to configure the nodes, the resources
and the flavors and refer to the Ironic documentation instead.

This change also removes the only reference in Sahara to the
old pxe_ipmitool driver, which is going to be removed:
http://lists.openstack.org/pipermail/openstack-dev/2018-March/128438.html

Remove also the reference to a bug which was fixed for a (long) while.

Change-Id: I5366298f82afb928b7c7520e2eb815d3ff597885