Replace the deleted keypair in clusters for API v2
This commit allows users to create a new keypair to access to
the running cluster when the cluster's keypair is deleted. But
the name of new keypair should be same as the deleted one, and
the new keypair is available for cluster scaling.
we can implement it via update_keypair parameter.

Change-Id: I3bc0ba52ceebff8ae128321aeb9fdbb5769baaf5
Closes-Bug:#1745142