Fixing extjs check on cdh and mapr
On images generated from sahara-image-elements cluster is being stuck
due to double execution of the unzip code.

This fix prevents the code to be executed twice and in the case it does we
force unzip to overwrite.

A release note was added to this stable/queens backport because the issue
affected the stable release 8.0.1.

Change-Id: I73836a516b839bdb368997af3693c139c1fff390
Story: #2002551
Task: #22113
(cherry picked from commit f5bf9ff2643a18cc91c2e62cfd086e869fd00b5d)