api-ref: move to a v1.1 sub-folder
Make space for v2.
It should not break too much existing links, only the links to
local anchors (like #clusters, #jobs, etc) will be broken,
because it is not possible to redirect server-side based on an
URL fragment like that.
Luckily the generated content ends all in the same HTML page,
so the old links will end up in the new top-level index
which points to the index for v1.1 API documentation (and later
for v2). Not too bad.

Change-Id: I9a75e6567798b8d6d3deed8fc66e362ad2d86b5a