Increase the startup time of ambari-server to 180s
It seems to be easy to hit timeout errors, so let's increase
the value of the timeout when ambari-server is started.
The option is recognized only from Ambari 2.5 onwards, but
the plugin uses Ambari 2.6 for all versions but HDP 2.3.

Story: 2004102
Task: 27509
Change-Id: I47c6dd14585ed20998fcf4a068cf1e7144eaf026
(cherry picked from commit 8bded6c9319370872581543325e3c98dcbedb0e1)