Migrate away from oslo_i18n.enable_lazy()
The method is known to be problematic, and it is going to be
deprecated:
http://lists.openstack.org/pipermail/openstack-operators/2018-November/016139.html

Moreover, its usage in Sahara produces some nasty side-effects
when Python 3.x is used. The provisioning of the clusters ends
abruptly when sahara-engine tries to update the status in the
database, with a failure inside oslo_i18n.

Change-Id: Icd2f22a0cb9e7642099600eeb38d3090d046c304
Story: 2004657
Task: 28613