Use venv-py2 to run sahara-scenario, remove the py3 job
This patch is required only on the stable branches, because
master will end up using Python 3 to execute sahara-scenario.
It ensures that sahara-scenario will always be executed using
Python 2 on this branch.

Also remove the Python 3 experimental job: it is not working
and it will be fixed only in master/future Stein. So save
a bit of resources.

Depends-On: https://review.openstack.org/608211
Change-Id: I41a26d3a7ca47c40b9fac815552d5a601f131bae