Fixing NTP issues for CDH plugin
First, we are fixing the adding of a new NTP server to ntp.conf.
The way it is currently done the server is added to the end of the
file and we fixed it to add it to the beginning of the config file.
Second, we are changing the order of when we set up ntp and the
service is started, because CDH fails to detect the sync'd time if
it has be synchronized after the start of the service.

Story: #2004981
Task: #29440

Change-Id: Ic2d7697a4036f5f689f065db081d53d4b37532cf
(cherry picked from commit 8f62975a8a84119c79ffdfa42d835665c7eeb515)