Add API v2 jobs (scenario, tempest); buildimages fixes
Run also: the API v2 scenario and tempest tests on both the check
and the gate queue.

Remove the list of plugins from required-project, as it was added
to the definition of the job in sahara-tests.git.

Finally, make sure that the base buildimages job depend on the
sahara repository and that it uses the sahara source directory
as working directory.

Change-Id: I97a302e253516cdd827c522178fca5fb5590f00c