grenade: re-enable, really test rocky->master
So far grenade has been testing a queens->rocky upgrade.

Implement the grenade code required to upgrade to the current
master. The plugin code installs the split plugins, which
need to be listed in the PROJECTS variable in order to be cloned.

Also, don't deploy sahara-all even when deploying rocky
(it was already deprecated back then).

Finally, deploying sahara-api using wsgi has been the default
for a while and for sure when the upgrade starts from rocky
there is no need to disable it.

The grenade test should be stable enough again, so re-enable it
in both the check and the gate queue.

Change-Id: I000aae6c093c32ab1c3505448e2b9ec52b0529a2