Update keystone_authtoken config reference
The auth_uri parameter of keystonemiddleware was renamed to
'www_authenticate_uri'[1], so update the documentation accordingly.
Additionally, the identity_uri parameter was deprecated in favor of
using a keystoneauth plugin[2], so this change updates the documentation
to point to the plugin parameters. Finally, the keystone v2 API was
dropped[3], so reference the v3 API.

[1] https://review.opendev.org/508522
[2] https://review.opendev.org/127066
[3] https://review.opendev.org/572846

Change-Id: I146e6d92008409ce219b18caaaa643c7f4496054