Update hacking for Python3
The repo is Python 3 now, so update hacking to version 3.0 which
supports Python 3.

Update local hacking checks for new flake8.

Note: The repo has been using an uncapped requirements, so was testing with
hacking 2.0 which disabled the old way of local testing integration.

Remove S367 and S368 since they are not working under python3
"NOTE(Kezar): This checks a good enough if we have only py2.7 supported.
 As soon as we'll get py3.x we need to drop it or rewrite. You can read more
 about it in dev-list archive, topic: "[hacking]proposed rules drop for 1.0"

Also, hacking now has H304 and H306 which test exactly the same.

Change-Id: I0b21554a5f9a69ed631b18c427b80c082011f774