Preinstall requirements for docs build
Sphinx installs the package with its dependencies, but it does not use
the constraints. This can lead to errors if a dependency introduces any
backward incompatible changes. To avoid this the dependencies needs to
be installed prior to sphinx run by tox with the correct constraints.

Conflicts:
    tox.ini

NOTE(elod.illes): conflict is due to branch specific upper constraints
setting patch.

NOTE(elod.illes): sahara-openstack-ansible-functional job is set as
non-voting as it seems there is a problem in openstack-ansible and the
job fails plus it is also set to non-voting in master until patch
Ib6e09ab83b277704a603c02f0927054478cef8a7 is merged. This is needed to
unblock the gate.

Change-Id: I4ba83f1007dd11801b8e32c6d7566769e005480b
(cherry picked from commit e0136189fc147deb7c4b92ed93eabb3cdcb9b7bb)