Drop tripleo centos-7 scen3 job definition - not used
This is needed by [1] where this job is being removed. Note
that this job as defined will only run for stable/stein [2].
Also note that this job doesn't appear to be running anywhere
it is not wired up in this repo [3] or anywhere else.

[1] https://review.opendev.org/c/openstack/tripleo-ci/+/838110/2#message-2efaca7aaf100874fd6996f18f0e74ae942149c0
[2] https://opendev.org/openstack/tripleo-ci/src/commit/4daed1b14c582b442f64ec45b1ddba5b58e03986/zuul.d/standalone-jobs.yaml#L476
[3] https://codesearch.opendev.org/?q=sahara-tripleo-ci-centos-7-scenario003-standalone&i=nope&literal=nope&files=&excludeFiles=&repos=

(cherry-picked from commit 4af54a062c3536ecce37e1386ee549fadbf3392d)

Change-Id: I7a4071e52f68fea7d07bad690f8590cbb3fb4fa4