Fix Flask 2.1 compatibility, disable unstable job
Since Flask 2.1, the same blueprint cannot be registered multiple
times with different url_prefix values, unless the 'name' parameter
is set. Otherwise the following error is reported: "The same blueprint
cannot be registered with the same name. Use name= when registering
to specify a unique name."

At the same time, disable the openstack-ansible job,
which started to fail again, but at this point the priority
is to unblock the gates.

Change-Id: I50df2fec3fd312082a14ef2a5b6774a2515afb44