Make ApiValidator to work with jsonschema>=4.0.0
As of Zed jsonschema version has been bumped to 4.14.0 in
upper-constraints, which means project should support running with
this version. This patch aims to fix ApiValidator that is not compatible
with new jsonschema.

Other deprecations, like FormatChecker.cls_checks decorator are left
intact and must be covered with follow-up patches.

We also move zuul queues to project scope as otherwise jobs do not
run due to config error.

We also replace iteritems with six method as iteritems have been dropped
from Python 3.

Change-Id: Ia8b69587aa9b3d04ffcdea7c6b97a8ae65f67534